Switch to line based parsing
We now parse the program line by line, and construct the blocks in the
"parser" module. This is uglier than parsing the program whole, but
has the huge advantage of much better error reporting, since the PEG
parsers are not very good at that, and with big blocks we usually get
the error at places much farther away than where they actually
happen.