Add FUNCTION block
For the function calls to work, we had to add a pass 3 to the
compiler. In the first pass, SUBs and FUNCTIONs are gathered. In the
second pass, we convert Lvalues matching functions to function
calls. And on the third, we check whether a function call is ever
being used as an Lvalue.