Handle variable auto-initialization
Decided that instead of initializing all variables in the QVM code,
the machine would perform auto-initialization. To do this, all READ
instructions are now typed. The machine would then initialize cells
upon first READ, based on the type of the READ instruction.

We also have a variant of READ instructions ending with @, which reads
a reference (these won't cause auto-initialization however, since we
don't want to have null references).