Refactor node
We don't have to implement "children" and "replace_child" methods in
all node sub-classes. Instead, implementing a "child_fields" class
attribute usually suffice.

child_fields should be a list of attribute names on the object. The
attributes should be the ones in the node containing either a child
node or a list of child nodes.

This covers almost all cases. In the case of the IfBlock, we needed
something like a list of list of children, so there the child_fields
attribute is a computed property, returning special names handled
properly by __getattr__ and __setattr__.