Add an empty line under error messages
We now add an empty line under error messages to make them more
distinct from the following lines of code.

For example, in the following:

 ||  DIM SHARED arena(1 TO 50, 1 TO 80) AS arenaType
 ||  DIM SHARED curLevel, colorTable(10)
 ||
 >>      RANDOMIZE TIMER
 ::      ^
 ::      Sub-program not found
 ||      GOSUB ClearKeyLocks
 ||      Intro
 ||      GetInputs NumPlayers, speed, diff$, monitor$

The error message and the GOSUB command have the same indentation,
causing the error messages to be a little indistinct. We now show this
as:

 ||  DIM SHARED arena(1 TO 50, 1 TO 80) AS arenaType
 ||  DIM SHARED curLevel, colorTable(10)
 ||
 >>      RANDOMIZE TIMER
 ::      ^
 ::      Sub-program not found
 ::
 ||      GOSUB ClearKeyLocks
 ||      Intro
 ||      GetInputs NumPlayers, speed, diff$, monitor$