Fix SUB/FUNCTION parameter type deduction
The types of the parameters to SUBs and FUNCTIONs should be deducated
based on the top-level DEF* statements. We gather DEF* statements in
pass 2, in order to avoid them affecting statements before them. This
however meant that they did not effect parameter types.

We now fix this by temporarily gathering top-level DEF* statements and
using them to deduce SUB/FUNCTION parameter type deduction.