Fix: codegen variables not accounted for in frame
The frame instructions generated were not accounting for the variables
that were being added later by codegen. We now fix this by allowing
instruction operands to be functions. The `local_vars_size` operand of
the frame instruction is now passed as a function, which will return
the updated result when being assembled.