Fix CONST code generation
We would encounter an error when calling a sub/function with a const
as an argument and no optimizations enabled. The reason was that we
were trying to generate code for pushing a reference to a constant.

We now check Lvalue nodes for referring to constants, and when they
do, replace the node with the constant value.

A clone method has been added to the Node class so we can copy
constant value nodes when replacing constant names with them, so each
can have a separate loc_start/loc_end pair.