Allow expression evaluation in dbg
The dbg PRINT command now accepts an expression, instead of a single
variable.

In order to support this, a significant amount of refactoring has been
done, so that the evaluation infrastructure is not dependent on the
presence of a compilation unit.

Some QVM related code has been moved away from the compiler to the qvm
package.

Evaluating user-defined types is not supported yet.