Distinguish between empty and left-out data items
An left-out empty data item like this:

   DATA

or the second item in this:

   DATA 1,,2

is slightly different than an empty quoted value like this:

   DATA ""

The former can be read both as a string and as a number (when read as
a number, it reads 0), while the latter cannot be read as a number.

We now use a one-value enum to represent a left-out empty value. We
could use None for this, but a named singleton (enum here) shows the
intention better.